Resolve unknown topic names on FetchRequest
We want to avoid asynchronously calling forward* methods on the KrpcFilterContext
because this could cause us to re-order messages and cause failures at the client
because they expect responses in the order they sent them.

Instead we will send a fire-and-forget metadata request to resolve unknown topic
ids when handling the FetchRequest, we do this before forwarding the FetchRequest
so the broker should respond to that first and we should know the mapping before
receiving the fetch response. If we do not know the topic name mapping when we
receive the response we will forward an error to the client immediately (preventing
problems with out of order messages) and sending another fire-and-forget request
for metadata.