Filter: load JSON definitions and policies from file
Why:
We want users to have access to the full Smorgasbord of configuration
including different encryption policies per-topic and support for
multiple KMS implementations.

With this change the Kroxylicious Filter will require configuration
to point to the two JSON files. This gives us the closest integration
with the encryption module because `JsonPolicyLoader.loadTopicPolicies`
works in terms of File objects and does some useful work like validating
and associating policies and KMS implementations.

In future, if we want to be able to embed the KMS definitions and
topic policies in the Filter configuration it might be nice if
`JsonPolicyLoader` had another method like:
`List<TopicPolicy> loadTopicPolicies(List<TopicPolicy> topicPolicies, Map<String, KmsDefinition> kmsDefs)`
that did the same work without the dependency on loading JSON from File.

Signed-off-by: Robert Young <robeyoun@redhat.com>