Fix bug with requesting topic metadata
Why:
At apiVersion 10, the MetadataRequest is only allowed to take null topic ids.
So it responds with an error when we set the id. We need to be at apiVersion
12. We were also missing a guard to prevent caching in this error condition
which allowed an empty topic name to be put in the cache on error. Resulting
in no decryption being applied.

Signed-off-by: Robert Young <robeyoun@redhat.com>