Enable dynamic reload of encryption config on config file modification
Users may now set a reloadOnFileChange boolean within the inMemoryPolicyRepository
configuration. This will instruct the Filter to watch both configuration files for
changes and to reload the encryption module.

Note that the EncryptionModule will be shared by all Channels that want to use
an inMemoryPolicyRepository backed by the same pair of files.

Why:
Users may wish to add a policy for a new topic before it is used, without
having to restart kroxylicious and kill all of it's filters.